requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
anyio==4.6.2
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
databases==0.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
logging==0.4.9.6
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.0
pyTelegramBotAPI==4.23
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
telegram==0.0.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
